requirements.txt
alembic==1.15.2
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dacite==1.9.2
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
Flask==3.0.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.1
h2o==3.44.0.3
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==7.2.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.41.1
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.8.2
mlflow==2.10.2
multimethod==1.10
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.58.1
numpy==1.25.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pandera==0.18.2
patsy==1.0.1
phik==0.12.4
pillow==11.2.1
pluggy==1.5.0
protobuf==4.25.7
puremagic==1.28
pyarrow==14.0.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.32.3
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.12.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
statsmodels==0.14.4
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
visions==0.7.5
waitress==2.1.2
websocket-client==1.8.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
ydata-profiling==4.6.4
zipp==3.21.0
